+++ inject/inject-thrift-client/src/test/scala/com/twitter/inject/thrift/FilteredThriftClientModuleNonMuxTest.scala
@@ -5,7 +5,7 @@ import com.twitter.inject.app.TestInjector
 import com.twitter.inject.modules.StatsReceiverModule
 import com.twitter.inject.thrift.filters.ThriftClientFilterBuilder
 import com.twitter.inject.thrift.modules.{FilteredThriftClientModule, ThriftClientIdModule}
-import com.twitter.inject.{Injector, InjectorModule, IntegrationTest, TypeUtils}
+import com.twitter.inject.{Injector, IntegrationTest, TypeUtils}
 import com.twitter.util.Future
 
 class FilteredThriftClientModuleNonMuxTest extends IntegrationTest {
@@ -14,13 +14,12 @@ class FilteredThriftClientModuleNonMuxTest extends IntegrationTest {
     modules = Seq(
       FilteredThriftClientModuleNonMux,
       ThriftClientIdModule,
-      StatsReceiverModule,
-      InjectorModule
+      StatsReceiverModule
     ),
     flags = Map("com.twitter.server.resolverMap" -> "greeter-thrift-service=nil!")
   ).create
 
-  lazy val greeter =
+  lazy val greeter: Greeter[Future] =
     injector.instance[Greeter[Future]](TypeUtils.asManifest[Greeter[Future]])
 
   test("bind expected type properly") {